Switched to MariaDB, now uses real barcode scanner instead of camera, exports to Google Sheets
Switched code from MySQL to MariaDB for Raspberry Pi support.

We have acquired an actual barcode scanner so I have rewritten the program to expect barcode data to be input via the terminal instead of through a camera with image processing.

Instead of exporting user data to a CSV file, attendance data is exported directly to Google Sheets